third_party/WebKit/LayoutTests/platform/linux/storage/websql/sql-error-codes-expected.txt

CONSOLE ERROR: line 37: Uncaught Exception thrown in transaction callback.
CONSOLE ERROR: line 69: Uncaught Cannot call toString() on this object.
This test tests the error codes reported in exceptional situations.
PASS: expected and got error code UNKNOWN_ERR
PASS: expected and got error code UNKNOWN_ERR
PASS: expected and got error code SYNTAX_ERR
PASS: expected and got error code SYNTAX_ERR
PASS: expected and got error code UNKNOWN_ERR
PASS: expected and got error code CONSTRAINT_ERR
PASS: expected and got error code QUOTA_ERR
FAIL: expected error code VERSION_ERR (2); got 1
